Change top-level namespace
This is not intended as stealing this excellent work. As the source
repository is now archived and no PR can be open against it, or merged
onto it, this renaming only aims at publishing a non-single-namespaced
version of this library, which can be used as a drop-in replacement.

In order to respect the authorship attribution, docstring and meta
have been added.